Create README.md
I have developed this project using Android with java. In this application I have used Xml for designing and java in backend and for database I have used firebase because it stores real time data in apps. This Chatting app is made for users to communicate with each other online.